refactor(enterprise): extract multi-replica subscribe into enterprise/coderd/chatd
Move NewMultiReplicaSubscribeFn and RemotePartsProvider into a
dedicated enterprise/coderd/chatd package. This gives the relay
merge logic its own package with proper unit tests while keeping
enterprise/coderd/chats.go focused on the HTTP relay plumbing.

- enterprise/coderd/chatd/chatd.go: RemotePartsProvider type and
  NewMultiReplicaSubscribeFn (all relay/pubsub merge logic)
- enterprise/coderd/chatd/chatd_test.go: relay reconnect, async
  non-blocking, snapshot delivery, and stale dial discard tests
- enterprise/coderd/chats.go: only newRemotePartsProvider and relay
  header helpers remain
- enterprise/coderd/chats_test.go: integration tests only
- Removed export_test.go (no longer needed)
