+++ enterprise/coderd/chats.go
@@ -10,6 +10,7 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/v2/codersdk"
+	entchatd "github.com/coder/coder/v2/enterprise/coderd/chatd"
 	"github.com/coder/websocket"
 )
 
@@ -21,21 +22,6 @@ const (
 	cookieHeader        = "Cookie"
 )
 
-// remotePartsProvider returns a snapshot and live stream of
-// message_part events from the replica that is running the chat.
-// Called when the chat is actively running on a different replica.
-type remotePartsProvider func(
-	ctx context.Context,
-	chatID uuid.UUID,
-	workerID uuid.UUID,
-	requestHeader http.Header,
-) (
-	snapshot []codersdk.ChatStreamEvent,
-	parts <-chan codersdk.ChatStreamEvent,
-	cancel func(),
-	err error,
-)
-
 // newRemotePartsProvider creates a remotePartsProvider that dials a remote
 // replica's stream endpoint to fetch message_part events. It filters to only
 // forward message_part events since durable events come via pubsub.
@@ -43,7 +29,7 @@ func newRemotePartsProvider(
 	resolveReplicaAddress func(context.Context, uuid.UUID) (string, bool),
 	replicaHTTPClient *http.Client,
 	replicaID uuid.UUID,
-) remotePartsProvider {
+) entchatd.RemotePartsProvider {
 	return func(
 		ctx context.Context,
 		chatID uuid.UUID,
