+++ enterprise/coderd/coderd.go
@@ -45,6 +45,7 @@ import (
 	agplusage "github.com/coder/coder/v2/coderd/usage"
 	"github.com/coder/coder/v2/coderd/wsbuilder"
 	"github.com/coder/coder/v2/codersdk"
+	entchatd "github.com/coder/coder/v2/enterprise/coderd/chatd"
 	"github.com/coder/coder/v2/enterprise/coderd/connectionlog"
 	"github.com/coder/coder/v2/enterprise/coderd/dbauthz"
 	"github.com/coder/coder/v2/enterprise/coderd/enidpsync"
@@ -204,7 +205,7 @@ func New(ctx context.Context, options *Options) (_ *API, err error) {
 	// Use a closure that captures api by reference so it can access
 	// api.AGPL.ID after coderd.New is called. The SubscribeFn is
 	// only invoked from Subscribe, which happens after init.
-	options.Options.ChatSubscribeFn = newMultiReplicaSubscribeFn(
+	options.Options.ChatSubscribeFn = entchatd.NewMultiReplicaSubscribeFn(
 		func(
 			ctx context.Context,
 			chatID uuid.UUID,
